// src/collection/TreesGenerator2D/types.ts
/**
 * @File   : types.ts
 * @Author : dtysky (dtysky@outlook.com)
 * @Date   : 2018-4-10 11:58:58
 * @Link: dtysky.moe
 */
export interface IControlOptions {
  clear: boolean;
  MAX_AGE: number;
  LEAF_DISTANCE: number;
  DIAMETER: number;
  LEAFE_SIZE: number;
  TREE_VARIANCE: number;
  BRANCH_VARIANCE: number;
  DRAW_DISTANCE: number;
}